Wed, Aug 31, 2022, 4:55 AM +00:00

+++ filter-less/helpers.c
@@ -59,10 +59,11 @@ void reflect(int height, int width, RGBTRIPLE image[height][width])
 {
     // nested for loop to iterate through each pixel of the image
     for (int i = 0; i < height; i++) {
+        
         for (int j = 0; j < width; j++) {
-            float red = image[i][j].rgbtRed;
-            float green = image[i][j].rgbtGreen;
-            float blue = image[i][j].rgbtBlue;
+            int red = image[i][j].rgbtRed;
+            int green = image[i][j].rgbtGreen;
+            int blue = image[i][j].rgbtBlue;
 
             int sepiaRed = round((red * 0.393) + (green * 0.769) + (blue * 0.189));
             int sepiaGreen = round((red * 0.349) + (green * 0.686) + (blue * 0.168));
